requirements.txt
absl-py==1.4.0
ai-edge-litert==1.2.0
anyio==4.7.0
apache-beam==2.63.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.7.1
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
avro-python3==1.10.2
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.1
crcmod==1.7
cycler==0.12.1
Cython==3.0.12
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
Django==4.2.20
django-bootstrap-v5==1.0.11
dm-tree==0.1.9
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
einops==0.8.1
etils==1.12.2
executing==2.1.0
fastavro==1.10.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.13.1
flatbuffers==25.2.10
fonttools==4.55.3
fqdn==1.5.1
fsspec==2024.2.0
gast==0.6.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.69.1
grpcio==1.65.5
h11==0.14.0
h5py==3.13.0
hdfs==2.7.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpickle==3.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server-mathjax==0.2.6
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.3.3
jupyterlab_git==0.50.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kaggle==1.7.4.2
keras==3.9.0
kiwisolver==1.4.7
libclang==18.1.1
lvis==0.5.3
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
nbclient==0.10.1
nbconvert==7.16.4
nbdime==4.0.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
object_detection @ file:///workspaces/TrashClass/models/research
objsize==0.7.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
portalocker==3.1.1
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==3.20.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pydot==1.4.2
Pygments==2.18.0
pymongo==4.11.2
pyparsing==2.4.7
python-dateutil==2.9.0.post0
python-json-logger==3.2.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
sacrebleu==2.2.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
seqeval==1.2.2
setuptools==75.6.0
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.1
tensorflow-datasets==4.9.8
tensorflow-hub==0.16.1
tensorflow-io==0.37.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-model-optimization==0.8.0
tensorflow-text==2.18.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf-models-official==2.18.0
tf-slim==1.1.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
torch==2.5.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
